+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The **patch** part is incremented if multiple releases happen the same month
 * openvpn: use its own minifirewall file instead of zzz-custom
 * postgresql: Use only :alnum: characters for nrpe's password
 * vrrpd: check minifirewall version even in check mode
+* Import upstream release 26.02 of evoacme
 
 ### Fixed